requirements.txt
playwright>=1.40.0
openai>=1.12.0
openpyxl>=3.1.2
python-docx>=1.1.0
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
rich>=13.7.0
apscheduler>=3.10.4
sqlite-utils>=3.36
docx2pdf>=0.1.8
thefuzz[speedup]>=0.22.1
google-api-python-client>=2.114.0
google-auth-oauthlib>=1.2.0
click>=8.1.7
matplotlib>=3.8.0
discord.py>=2.3.2
cloudscraper>=1.2.71
tzdata>=2024.1
pytest>=8.0.0
